+++ CHANGELOG.md
@@ -9,6 +9,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Features
 
+- program: auction order params account for twap divergence ([#1882](https://github.com/drift-labs/protocol-v2/pull/1882))
+- program: add delegate stake if ([#1859](https://github.com/drift-labs/protocol-v2/pull/1859))
+
+### Fixes
+
+### Breaking
+
+## [2.138.0] - 2025-09-22
+
+### Features
+
+- program: support scaled ui extension ([#1894](https://github.com/drift-labs/protocol-v2/pull/1894))
+- Revert "Crispeaney/revert swift max margin ratio ([#1877](https://github.com/drift-labs/protocol-v2/pull/1877))
+
 ### Fixes
 
 ### Breaking
@@ -20,6 +34,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - program: post only respects reduce only ([#1878](https://github.com/drift-labs/protocol-v2/pull/1878))
 - program: add sequence id to exchange/mm oracle ([#1834](https://github.com/drift-labs/protocol-v2/pull/1834))
 - program: perp position max margin ratio ([#1847](https://github.com/drift-labs/protocol-v2/pull/1847))
+- program: add padding to swift messages ([#1845](https://github.com/drift-labs/protocol-v2/pull/1845))
 - program: rm lp ([#1755](https://github.com/drift-labs/protocol-v2/pull/1755))
 
 ### Fixes